modularized backend + updated .gitignore
