+++ backend/config/transporterConfig.js
@@ -1,0 +1,11 @@
+const nodemailer = require("nodemailer");
+
+module.exports = nodemailer.createTransport({
+    host: "smtp-mail.outlook.com",
+    port: 587,
+    secure: false, // true for 465, false for other ports
+    auth: {
+        user: process.env.EMAIL_USERNAME,
+        pass: process.env.EMAIL_PASSWORD,
+    },
+});
